Revert: correct Woods-Saxon parameters for finite nucleon size.
When calculating nuclear thickness functions using finite-size nucleons, the
nucleon profile effectively "smears out" the underlying distribution of nucleon
centers.  Therefore if nucleon positions are sampled from a Woods-Saxon
distribution, the resulting nuclear thickness will effectively have a somewhat
different (wider) distribution.

This method deconvolves the Gaussian nucleon profile from the Woods-Saxon
distribution in order to correct for the smearing effect, which leads to a
decreases in the total nuclear cross section (relative to no correction).  In
fact, the decrease is significant, roughly 7.7 down to 7.3 barn for 2.76 TeV
Pb-Pb and 0.4 fm nucleons.  Additionally, with no correction the total cross
section _increases_ as nucleon width increases, but with the correction the
cross section _decreases_ with nucleon width.  Given that the experimental cross
section is ~7.65 barn, this is somewhat concerning.

So while a correction like this can make sense, it probably shouldn't depend on
nucleon width.  The nucleon width shouldn't be taken this literally -- it's more
of an effective parameter intended to account for the increase of nucleon cross
section with energy.  Within the context of the model, increasing the nucleon
width _should_ increase the total nuclear cross section.

On a related note, there is some evidence that heavy nuclei have a "neutron
skin", i.e. the radial neutron distribution is extends somewhat further than
protons.  Since Woods-Saxon parameters are derived from electron scattering
experiments, they may neglect the neutron skin and hence underestimate the
nuclear radius.

Conveniently, it appears that these two effects roughly cancel: the finite-size
nucleon smearing increases the radius by about the same amount as a neutron
skin.  Given this and the good agreement of the model's Pb-Pb cross section with
experiment, we choose to leave the Woods-Saxon parameters unmodified.